rename decorateWithMetrics to metering
