+++ lib/Onyx.js
@@ -757,18 +757,18 @@ const Onyx = {
  */
 function applyDecorators() {
     // We're requiring the script dynamically here so that it's only evaluated when decorators are used
-    const decorate = require('./decorateWithMetrics');
+    const metering = require('./metering');
 
     // Re-assign with decorated functions
     /* eslint-disable no-func-assign */
-    get = decorate.decorateWithMetrics(get, 'Onyx:get');
-    set = decorate.decorateWithMetrics(set, 'Onyx:set');
-    multiSet = decorate.decorateWithMetrics(multiSet, 'Onyx:multiSet');
-    clear = decorate.decorateWithMetrics(clear, 'Onyx:clear');
-    merge = decorate.decorateWithMetrics(merge, 'Onyx:merge');
-    mergeCollection = decorate.decorateWithMetrics(mergeCollection, 'Onyx:mergeCollection');
-    getAllKeys = decorate.decorateWithMetrics(getAllKeys, 'Onyx:getAllKeys');
-    initializeWithDefaultKeyStates = decorate.decorateWithMetrics(initializeWithDefaultKeyStates, 'Onyx:defaults');
+    get = metering.decorateWithMetrics(get, 'Onyx:get');
+    set = metering.decorateWithMetrics(set, 'Onyx:set');
+    multiSet = metering.decorateWithMetrics(multiSet, 'Onyx:multiSet');
+    clear = metering.decorateWithMetrics(clear, 'Onyx:clear');
+    merge = metering.decorateWithMetrics(merge, 'Onyx:merge');
+    mergeCollection = metering.decorateWithMetrics(mergeCollection, 'Onyx:mergeCollection');
+    getAllKeys = metering.decorateWithMetrics(getAllKeys, 'Onyx:getAllKeys');
+    initializeWithDefaultKeyStates = metering.decorateWithMetrics(initializeWithDefaultKeyStates, 'Onyx:defaults');
     /* eslint-enable */
 
     // Re-expose decorated methods
@@ -779,9 +779,9 @@ function applyDecorators() {
     Onyx.mergeCollection = mergeCollection;
 
     // Expose stats methods on Onyx
-    Onyx.getMetrics = decorate.getMetrics;
-    Onyx.resetMetrics = decorate.resetMetrics;
-    Onyx.printMetrics = decorate.printMetrics;
+    Onyx.getMetrics = metering.getMetrics;
+    Onyx.resetMetrics = metering.resetMetrics;
+    Onyx.printMetrics = metering.printMetrics;
 }
 
 export default Onyx;
