+++ lib/metering.js
